Ajoute les champs des propriétés de l'utilisateur dans la page d'édition
