+++ public/modules/interactions/brancheMiseEnAvantSaisie.js
@@ -1,0 +1,4 @@
+const brancheMiseEnAvantSaisie = () => $("input[type!='radio'][type!='checkbox'], select")
+  .on('change', (evenement) => $(evenement.target).toggleClass('champ-saisi', evenement.target.value !== ''));
+
+export default brancheMiseEnAvantSaisie;
