+++ src/vues/fragments/utilisateur/nouveauMotDePasse.pug
@@ -1,0 +1,5 @@
+mixin nouveauMotDePasse
+  h2 Changer de mot de passe
+  label Nouveau mot de passe
+    .infos-complementaires (Si vous laissez ce champ vide, le mot de passe ne sera pas modifié.)
+    input(id='mot-de-passe', name = 'motDePasse', type='password')
